修复pg初始化

+++ pages/api/admin/init.ts
@@ -13,10 +13,10 @@ export default async function (req: VercelRequest, res: VercelResponse) {
     await db_bitio_pool
       .query(
         `CREATE TABLE blacklist(
-      uid INT PRIMARY KEY     NOT NULL,
-      type            INT     NOT NULL,
+      uid INT8 PRIMARY KEY     NOT NULL,
+      type            INT8     NOT NULL,
       reason          TEXT,
-      ban_until       INT
+      ban_until       INT8
     )`
       )
       .catch((err) => console.error(err));
@@ -24,9 +24,9 @@ export default async function (req: VercelRequest, res: VercelResponse) {
       .query(
         `CREATE TABLE cache(
       cid INT PRIMARY KEY,
-      ep              INT,
-      need_vip        INT,
-      exp             INT,
+      ep              INT8,
+      need_vip        INT8,
+      exp             INT8,
       data            JSON
     )`
       )
@@ -35,10 +35,10 @@ export default async function (req: VercelRequest, res: VercelResponse) {
       .query(
         `CREATE TABLE log(
       uid INT PRIMARY KEY,
-      vip_type        INT,
+      vip_type        INT8,
       access_key      TEXT,
       url             TEXT,
-      visit_time      INT
+      visit_time      INT8
     )`
       )
       .catch((err) => console.error(err));
